+++ changelog.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Inherits matcher not matching for inherited base type
 - IsMatcher now has human readable string representation
+- implements and inherits matcher now uses the schema that is documented
 
 ## [0.2] - 2023-08-04
 ### Added